Fix masking strategy tests that were broken.
We have several tests that verify that the actual output of a masking strategy that uses
a random number generator matches an exact expected output. This is not great because the
functions being tested never intended to provide a contract that they will always return
that exact output for a given random seed. This means that minor changes to the function
easily break the tests, making it extremely hard to maintain.

I have updated a few tests that were failing to instead test in a way that respects the
contract being provided without diving low-level to the extent that they rely on a specific
random seed. For example, if we use random masking with 50% encode 50% decode, then we can
verify that roughly 50% of the tokens are set ONLINE_ENCODER and 50% DECODER.